+++ Gemfile
@@ -1,11 +1,12 @@
 
 source 'https://rubygems.org'
 
-gem 'nanoc', '~> 4.12'  # engine
+gem 'nanoc', '~> 4.13'  # engine
 gem 'kramdown'          # markdown converter
 gem 'typogruby'         # text enhancements
 gem 'rouge'             # syntax highlighting
 gem 'slim'              # templates
+gem 'erubi'             # templates
 gem 'builder'           # xml sitemaps
 gem 'terser'            # js uglifying
 gem 'ruby-xxHash'       # for cache busting
@@ -13,11 +14,11 @@ gem 'ruby-xxHash'       # for cache busting
 group :nanoc do
   gem 'nanoc-dart-sass' # css conversion
   gem 'nanoc-live'      # automatic rebuilds
+  gem 'nanoc-external'  # tidy filter
 end
 
 group :development do
-  gem 'adsf'            # tiny webserver
-  gem 'puma'            # webserver backend
+  gem 'falcon'          # webserver backend
 end
 
 group :test do
@@ -26,4 +27,4 @@ group :test do
 end
 
 # nanoc is incompatible with ancient ruby
-ruby '>= 2.7'
+ruby '>= 3.2'
