--- content/ads.txt
@@ -1,1 +1,0 @@
-# We hate ads, so nothing to see here... Move along!
