+++ content/robots.txt
@@ -3,4 +3,4 @@ Disallow: /irc/log/
 Disallow: /rtp/
 Disallow: /player/doxydoc/
 
-Sitemap: https://easyrpg.org/sitemap.xml
+Sitemap: <%= @config[:base_url] %>/sitemap.xml
