+++ content/sitemap.xml
@@ -1,0 +1,1 @@
+<%= xml_sitemap(items: content_items()) %>
