+++ lib/sitemap.rb
@@ -8,8 +8,3 @@ def content_items
     !item[:is_hidden] && (item[:extension] == '.html' || item.path.to_s.end_with?('/'))
   end
 end
-
-# called by preprocess block
-def write_sitemap_file
-  @items.create("<%= xml_sitemap(items: content_items) %>", {}, '/sitemap.xml')
-end
