+++ nanoc.yaml
@@ -1,18 +1,15 @@
 # A list of file extensions that Nanoc will consider to be textual rather than
 # binary. If an item with an extension not in this list is found,  the file
 # will be considered as binary.
-text_extensions: [ 'adoc', 'asciidoc', 'atom', 'coffee', 'css', 'erb', 'haml', 'handlebars', 'hb', 'htm', 'html', 'js', 'less', 'markdown', 'md', 'ms', 'mustache', 'php', 'rb', 'rdoc', 'sass', 'scss', 'slim', 'tex', 'txt', 'xhtml', 'xml' ]
+text_extensions: [ 'adoc', 'asciidoc', 'atom', 'coffee', 'css', 'erb', 'haml',
+                   'handlebars', 'hb', 'htm', 'html', 'js', 'less', 'markdown',
+                   'md', 'ms', 'mustache', 'php', 'rb', 'rdoc', 'sass', 'scss',
+                   'slim', 'tex', 'txt', 'xhtml', 'xml' ]
 
 # The path to the directory where all generated files will be written to. This
 # can be an absolute path starting with a slash, but it can also be path
 # relative to the site directory.
 output_dir: output
-
-# A list of index filenames, i.e. names of files that will be served by a web
-# server when a directory is requested. Usually, index files are named
-# “index.html”, but depending on the web server, this may be something else,
-# such as “default.htm”. This list is used by Nanoc to generate pretty URLs.
-index_filenames: [ 'index.html' ]
 
 # Whether or not to generate a diff of the compiled content when compiling a
 # site. The diff will contain the differences between the compiled content
@@ -23,33 +20,6 @@ prune:
   # Whether to automatically remove files not managed by Nanoc from the output
   # directory.
   auto_prune: true
-
-  # Which files and directories you want to exclude from pruning. If you version
-  # your output directory, you should probably exclude VCS directories such as
-  # .git, .svn etc.
-  exclude: [ '.git' ]
-
-# The data sources where Nanoc loads its data from. This is an array of
-# hashes; each array element represents a single data source. By default,
-# there is only a single data source that reads data from the “content/” and
-# “layout/” directories in the site directory.
-data_sources:
-  - type: filesystem
-
-    # The path where items should be mounted (comparable to mount points in
-    # Unix-like systems). This is “/” by default, meaning that items will have
-    # “/” prefixed to their identifiers. If the items root were “/en/”
-    # instead, an item at content/about.html would have an identifier of
-    # “/en/about/” instead of just “/about/”.
-    items_root: /
-
-    # The path where layouts should be mounted. The layouts root behaves the
-    # same as the items root, but applies to layouts rather than items.
-    layouts_root: /
-
-    # The encoding to use for input files. If your input files are not in
-    # UTF-8 (which they should be!), change this.
-    encoding: utf-8
 
 # Configuration for the “check” command, which run unit tests on the site.
 checks:
@@ -90,4 +60,5 @@ environments:
     prod: true
 #  pr:
   netlify:
+    # url placeholder will be overwritten by build script
     base_url: DEPLOY_PRIME_URL
